+++ examples/protocol_test/src/bin/connector.rs
@@ -4,11 +4,13 @@ use fe2o3_amqp::{types::primitives::Value, Connection, Delivery, Receiver, Sende
 use tracing::{instrument, Level};
 use tracing_subscriber::FmtSubscriber;
 
+// const SASL_PLAIN: &str = "";
+const SASL_PLAIN: &str = "guest:guest";
 const BASE_ADDR: &str = "localhost:5672";
 
 #[instrument]
 async fn client_main() {
-    let url = format!("amqp://{}", BASE_ADDR);
+    let url = format!("amqp://{}@{}", SASL_PLAIN, BASE_ADDR);
     let mut connection = Connection::open("connection-1", &url[..]).await.unwrap();
 
     let mut session = Session::begin(&mut connection).await.unwrap();
@@ -40,7 +42,7 @@ async fn client_main() {
 #[tokio::main]
 async fn main() {
     let subscriber = FmtSubscriber::builder()
-        .with_max_level(Level::TRACE)
+        .with_max_level(Level::INFO)
         .finish();
     tracing::subscriber::set_global_default(subscriber).expect("setting default subscriber failed");