+++ fe2o3-amqp-types/src/performatives/attach.rs
@@ -2,7 +2,7 @@ use std::collections::BTreeMap;
 
 use serde_amqp::{
     macros::{DeserializeComposite, SerializeComposite},
-    primitives::{Boolean, Symbol, ULong},
+    primitives::{Boolean, Symbol, ULong, Array},
 };
 
 use crate::{
@@ -106,10 +106,10 @@ pub struct Attach {
     pub max_message_size: Option<ULong>,
 
     /// <field name="offered-capabilities" type="symbol" multiple="true"/>
-    pub offered_capabilities: Option<Vec<Symbol>>,
+    pub offered_capabilities: Option<Array<Symbol>>,
 
     /// <field name="desired-capabilities" type="symbol" multiple="true"/>
-    pub desired_capabilities: Option<Vec<Symbol>>,
+    pub desired_capabilities: Option<Array<Symbol>>,
 
     /// <field name="properties" type="fields"/>
     pub properties: Option<Fields>,