+++ fe2o3-amqp-types/src/sasl/mod.rs
@@ -1,10 +1,12 @@
 //! Types defined in AMQP 1.0 specification Part 5.3: SASL
 
 use serde_amqp::{
-    primitives::{Binary, Symbol},
+    primitives::{Binary, Symbol, Array},
     DeserializeComposite, SerializeComposite,
 };
 use serde_repr::{Deserialize_repr, Serialize_repr};
+
+mod mechanisms;
 
 /// 5.3.3.1 SASL Mechanisms
 /// Advertise available sasl mechanisms.
@@ -13,13 +15,11 @@ use serde_repr::{Deserialize_repr, Serialize_repr};
 ///     <field name="sasl-server-mechanisms" type="symbol" multiple="true" mandatory="true"/>
 /// </type>
 /// Advertises the available SASL mechanisms that can be used for authentication.
-#[derive(Debug, Clone, SerializeComposite, DeserializeComposite)]
-#[amqp_contract(
-    name = "amqp:sasl-mechanisms:list",
-    code = 0x0000_0000_0000_0040,
-    encoding = "list",
-    rename_all = "kebab-case"
-)]
+/// 
+/// NOTE: Serialize and Deserialize are manually implemented because 
+/// > A field which is defined as both multiple and mandatory MUST contain at least one value 
+/// (i.e. for such a field both null and an array with no entries are invalid).
+#[derive(Debug, Clone)]
 pub struct SaslMechanisms {
     /// sasl-server-mechanisms supported sasl mechanisms
     ///
@@ -28,7 +28,7 @@ pub struct SaslMechanisms {
     /// authenticate with it, then it SHOULD send a list of one element with its value as the
     /// SASL mechanism ANONYMOUS. The server mechanisms are ordered in decreasing level of
     /// preference.
-    pub sasl_server_mechanisms: Vec<Symbol>,
+    pub sasl_server_mechanisms: Array<Symbol>,
 }
 
 /// 5.3.3.2 SASL Init