+++ fe2o3-amqp/src/transport/protocol_header.rs
@@ -1,6 +1,14 @@
 //! Implements the protocol headers
 
-use std::convert::{TryFrom, TryInto};
+use std::{
+    convert::{TryFrom, TryInto},
+    io,
+};
+
+use bytes::{Buf, BufMut, Bytes};
+use tokio_util::codec::{Decoder, Encoder};
+
+use super::error::NegotiationError;
 
 const PROTOCOL_HEADER_PREFIX: &[u8; 4] = b"AMQP";
 
@@ -109,6 +117,13 @@ impl From<ProtocolHeader> for [u8; 8] {
     }
 }
 
+impl From<ProtocolHeader> for Bytes {
+    fn from(header: ProtocolHeader) -> Self {
+        let bytes: [u8; 8] = header.into();
+        Bytes::copy_from_slice(&bytes[..])
+    }
+}
+
 impl TryFrom<[u8; 8]> for ProtocolHeader {
     type Error = [u8; 8];
 
@@ -122,6 +137,45 @@ impl TryFrom<[u8; 8]> for ProtocolHeader {
         };
 
         Ok(Self::new(id, v[5], v[6], v[7]))
+    }
+}
+
+impl<'a> TryFrom<&'a [u8]> for ProtocolHeader {
+    type Error = &'a [u8];
+
+    fn try_from(value: &'a [u8]) -> Result<Self, Self::Error> {
+        if value.len() != 8 {
+            return Err(value);
+        }
+
+        if value[..4] != b"AMQP"[..] {
+            return Err(value);
+        }
+
+        let id = value[4].try_into().map_err(|_| value)?;
+
+        Ok(Self::new(id, value[5], value[6], value[7]))
+    }
+}
+
+impl TryFrom<Bytes> for ProtocolHeader {
+    type Error = Bytes;
+
+    fn try_from(value: Bytes) -> Result<Self, Self::Error> {
+        if value.len() != 8 {
+            return Err(value);
+        }
+
+        if value[..4] != b"AMQP"[..] {
+            return Err(value);
+        }
+
+        let id = match value[4].try_into() {
+            Ok(id) => id,
+            Err(_) => return Err(value),
+        };
+
+        Ok(Self::new(id, value[5], value[6], value[7]))
     }
 }
 
@@ -149,5 +203,49 @@ impl TryFrom<u8> for ProtocolId {
             _ => return Err(value),
         };
         Ok(val)
+    }
+}
+
+/// Encoder and Decoder for protocol headers
+#[derive(Debug, Clone)]
+pub struct ProtocolHeaderCodec {}
+
+impl ProtocolHeaderCodec {
+    /// Creates a new protocol header codec
+    pub fn new() -> Self {
+        Self {}
+    }
+}
+
+impl Encoder<ProtocolHeader> for ProtocolHeaderCodec {
+    type Error = io::Error;
+
+    fn encode(
+        &mut self,
+        item: ProtocolHeader,
+        dst: &mut bytes::BytesMut,
+    ) -> Result<(), Self::Error> {
+        let buf: [u8; 8] = item.into();
+        dst.put(&buf[..]);
+        Ok(())
+    }
+}
+
+impl Decoder for ProtocolHeaderCodec {
+    type Item = ProtocolHeader;
+    type Error = NegotiationError;
+
+    fn decode(&mut self, src: &mut bytes::BytesMut) -> Result<Option<Self::Item>, Self::Error> {
+        // If the bytes look valid, but a frame isn’t fully available yet, then Ok(None) is
+        // returned. This indicates to the Framed instance that it needs to read some more bytes
+        // before calling this method again.
+        if src.remaining() < 8 {
+            return Ok(None);
+        }
+
+        let bytes = src.split_to(8).freeze();
+        ProtocolHeader::try_from(bytes)
+            .map(Some)
+            .map_err(|b| NegotiationError::ProtocolHeaderMismatch(b))
     }
 }