+++ serde_amqp/Changelog.md
@@ -1,5 +1,19 @@
 # Change Log
 
+## 0.0.8
+
+1. Updated dependencies
+   1. `ordered-float` to "3"
+   2. `uuid` to "1"
+
+## 0.0.7
+
+1. Added `FromIterator` impl for `Array`
+
+## 0.0.6
+
+1. (forgot)
+
 ## 0.0.5
 
 1. Made `serde_amqp::read::{IoReader, SliceReader}` public.